% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{columnGraph}
\alias{columnGraph}
\title{Create a Column/Bar Graph from Data}
\usage{
columnGraph(sdata, xname = FALSE, yname = FALSE, tname = "", fname = "")
}
\arguments{
\item{sdata}{A data frame or tibble containing the data for the graph. The first column is used for the x-axis 
and the second column is used for the y-axis.}

\item{xname}{A character string to specify the label for the x-axis. If not provided (default is `FALSE`), the 
label will be the name of the first column.}

\item{yname}{A character string to specify the label for the y-axis. If not provided (default is `FALSE`), the 
label will be the name of the second column.}

\item{tname}{A character string to specify the title of the plot. If not provided (default is `""`), the title 
will be generated using the names of the first two columns, formatted as "Column1 / Column2 Analysis".}

\item{fname}{A character string to specify the name of the save graph file.}
}
\value{
The plot that was created.
}
\description{
This function generates a column (bar) graph using `ggplot2` based on the provided data frame. The function 
allows dynamic selection of the x and y axes by using the first two columns in the data frame, or custom axis 
labels and title. The plot is then saved as a PNG image using the `saveGraph` function. Saves the generated column/bar graph as a PNG file in the "reports" directory. The file is named using the 
  plot's title and includes a random string to ensure uniqueness.
}
\examples{
# Example data frame for creating a column graph
df <- tibble(Category = c("A", "B", "C"), Value = c(10, 20, 30))
columnGraph(df, xname = "Category", yname = "Value", tname = "Category Analysis")

# Example of creating a graph with default labels and title
columnGraph(df)

}
