% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryAnalysis.R
\name{countrySale}
\alias{countrySale}
\title{Summarize and Sort Sales or Returns by Country}
\usage{
countrySale(data, num, decreasing = TRUE, returns = FALSE)
}
\arguments{
\item{data}{A DataFrame or tibble containing the sales or return data, with columns for `Country` and `Quantity`.}

\item{num}{An integer representing the number of top or bottom countries to return based on the total quantity. 
If `num` is negative, the bottom `abs(num)` countries will be returned.}

\item{decreasing}{A logical value indicating whether to sort the result in descending order (`TRUE`, default) 
or ascending order (`FALSE`).}

\item{returns}{A logical value indicating whether to filter for returns (`TRUE`) or sales (`FALSE`). Defaults to `FALSE` (sales).}
}
\value{
A DataFrame or tibble with the top or bottom `num` countries sorted by total quantity, with the column name 
  reflecting either "TotalReturns" or "TotalSales" based on the `returns` argument.
}
\description{
This function summarizes the sales or returns data by country and sorts the total quantity in ascending or 
descending order. It uses the `saleOrReturn` function to filter data based on the `returns` argument and 
groups the results by country, summing the absolute quantity.
}
\details{
If `num` is negative, the function returns the bottom `abs(num)` countries based on total quantity in the 
opposite order (ascending if it was descending, and vice versa).
}
\examples{
df <- tibble(Country = c("UK", "USA", "UK", "USA"),
             Quantity = c(10, -5, 20, -2))
# Get top 2 countries based on total sales
country_sale_result <- countrySale(df, num = 2, returns = FALSE)
print(country_sale_result)

# Get bottom 2 countries based on total sales
country_sale_result <- countrySale(df, num = -2, returns = FALSE)
print(country_sale_result)

}
