% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data from CSV File}
\usage{
loadData(file_name = NULL)
}
\arguments{
\item{filename}{A character string representing the path to the CSV file to be loaded.}
}
\value{
A tibble or data frame containing the data from the CSV file.
}
\description{
This function loads a CSV file into a data frame or tibble. It automatically reads
the file and returns the contents.
}
\details{
The function uses `read.csv()` or `readr::read_csv()` to load the CSV file and 
automatically handles common data loading scenarios such as missing values and 
incorrect data types.
}
\examples{
# Load data from a CSV file
data <- loadData("path/to/your/file.csv")
# View the loaded data
head(data)

}
