% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{saleOrReturn}
\alias{saleOrReturn}
\title{Filter Sales or Returns Data Based on Quantity}
\usage{
saleOrReturn(data, returns = FALSE, name1 = "Returns", name2 = "Sales")
}
\arguments{
\item{data}{A DataFrame or tibble containing the sales or return data.}

\item{returns}{A logical value indicating whether to filter for returns (`TRUE`) or sales (`FALSE`).
Defaults to `FALSE` (sales).}

\item{name1}{A string representing the name to use for returns, defaults to "Returns".}

\item{name2}{A string representing the name to use for sales, defaults to "Sales".}
}
\value{
A list with two elements:
  \describe{
    \item{data}{A DataFrame or tibble filtered based on quantity (greater than 0 for sales, less than 0 for returns).}
    \item{rname}{A string representing the name of the category ("Returns" or "Sales").}
  }
}
\description{
This function filters the data based on the quantity. It selects rows with a quantity greater than 0 for sales 
or a quantity less than 0 for returns, depending on the value of the `returns` argument. The function also 
assigns a column name based on the specified `name1` and `name2` arguments.
}
\examples{
df <- tibble(InvoiceNo = c(1, 2, 3, 4), Quantity = c(10, -5, 3, -2))
result <- saleOrReturn(df, returns = TRUE)
print(result$data)  # Rows where Quantity < 0 (returns)
print(result$rname)  # "Returns"

}
