% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{saveGraph}
\alias{saveGraph}
\title{Save Plot as an Image File}
\usage{
saveGraph(p, fname = "")
}
\arguments{
\item{p}{A ggplot2 plot object to be saved.}

\item{fname}{A character string specifying the desired filename (without the file extension). 
If not provided (default is `""`), the function will use the plot's title as the filename.}
}
\value{
Location of saved plot.
}
\description{
This function saves a given ggplot2 plot as an image file in the "reports" directory. If no name is provided, 
the function will use the title of the plot as the default filename, with any spaces replaced by underscores 
and slashes replaced with "by". A random string is appended to the filename to ensure uniqueness.
}
\details{
If the "reports" directory does not exist, it will be created.
}
\examples{
# Example of saving a plot with a custom filename
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point() + ggtitle("Car Weight vs MPG")
saveGraph(p, "custom_plot_name")

# Example of saving a plot with a default filename (based on the plot's title)
saveGraph(p)

}
