% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{sortDataframe}
\alias{sortDataframe}
\title{Sort DataFrame by a Given Column}
\usage{
sortDataframe(data, column, decreasing = TRUE)
}
\arguments{
\item{data}{A DataFrame or tibble that you want to sort.}

\item{column}{A character string representing the column name to sort by.}

\item{decreasing}{A logical value. If `TRUE` (default), the data will be sorted in descending order, 
otherwise, in ascending order.}
}
\value{
A sorted DataFrame or tibble based on the given column.
}
\description{
This function sorts a data frame based on a specified column, in either ascending or descending order.
}
\examples{
df <- tibble(StockCode = c("A1", "B2", "A1", "B2"),
             Amount = c(100, -50, 200, -30))
sorted_df <- sortDataframe(df, "Amount", decreasing = TRUE)
print(sorted_df)

}
